+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update Vue to v3.5.3
 - Replace placeholders in preview urls (id, locale, field value)
 - Mark unpublished linked content items
+- Pass a list of checks to perform on system healthcheck route (db, memory, fs, redis, smtp, custom)
 
 ## 2.9.1 (2024-08-26)